requirements.txt
streamlit
opencv-python
ultralytics
Pillow